refactor(engines): split into ecore / graph / smartform sub-packages
Move all engine code from the flat internal/engines/ into three sub-packages
to group files by domain:

  internal/engines/ecore/     — shared runtime config + HTTP helpers
    runtime.go                  Config, Cfg, Configure, SetBaseURL, ResetAuth,
                                AuthHeader, BuildBaseURL, EnsureAuth
    httpclient.go               APIHTTPClient (shared TLS-aware HTTP client)
    helpers.go                  Seg, RequireUUID, PapiGET, PapiPOST, PapiPUT
    ecore_test.go               TestRequireUUID, TestSeg, TestBuildBaseURL*

  internal/engines/graph/     — graph sync, layout, pruning, upload, chart
    sync.go, push.go, compact_layout.go, placements.go, prune_edges.go,
    chart.go, upload.go, svg.go, helpers.go, register.go
    sync_test.go, security_test.go

  internal/engines/smartform/ — CDU create/pull/push, releases, file history
    create.go, pull.go, push.go, releases.go, file_history.go, register.go

  internal/engines/           — thin wiring layer
    register.go   calls graph.Register + smartform.Register
    runtime.go    re-exports Configure/SetBaseURL/ResetAuth from ecore
                  (so cmd/server and tools/build.go need no import changes)

All tests pass; go build and go vet clean.
