+++ plugins/simulator/mcp-server/internal/engines/graph/placements.go
@@ -1,4 +1,4 @@
-package engines
+package graph
 
 import (
 	"context"
@@ -8,6 +8,7 @@ import (
 	"net/http"
 	"os"
 
+	"github.com/corezoid/simulator-ai-plugin/plugins/simulator/mcp-server/internal/engines/ecore"
 	"github.com/mark3labs/mcp-go/mcp"
 )
 
@@ -43,7 +44,7 @@ type paginatedLayerActor struct {
 }
 
 func handleGetAllLayerPlacements(ctx context.Context, req mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-	if authResult := ensureAuth(ctx); authResult != nil {
+	if authResult := ecore.EnsureAuth(ctx); authResult != nil {
 		return authResult, nil
 	}
 	args := req.GetArguments()
@@ -51,23 +52,23 @@ func handleGetAllLayerPlacements(ctx context.Context, req mcp.CallToolRequest) (
 	if layerID == "" {
 		return mcp.NewToolResultError("[Error] layerId is required"), nil
 	}
-	if r := requireUUID("layerId", layerID); r != nil {
+	if r := ecore.RequireUUID("layerId", layerID); r != nil {
 		return r, nil
 	}
 
-	client := apiHTTPClient()
+	client := ecore.APIHTTPClient()
 
 	var rows []layerPlacementRow
 	const limit = 100
 	offset := 0
 	for {
 		u := fmt.Sprintf("%s/graph_layers/paginated/%s?type=nodes&limit=%d&offset=%d",
-			buildBaseURL(), layerID, limit, offset)
+			ecore.BuildBaseURL(), layerID, limit, offset)
 		httpReq, err := http.NewRequestWithContext(ctx, "GET", u, nil)
 		if err != nil {
 			return mcp.NewToolResultError(fmt.Sprintf("[Error] new request: %v", err)), nil
 		}
-		httpReq.Header.Set("Authorization", authHeader())
+		httpReq.Header.Set("Authorization", ecore.AuthHeader())
 		resp, err := client.Do(httpReq)
 		if err != nil {
 			return mcp.NewToolResultError(fmt.Sprintf("[Error] http: %v", err)), nil
