+++ plugins/simulator/mcp-server/internal/engines/graph/svg.go
@@ -1,4 +1,4 @@
-package engines
+package graph
 
 import (
 	"bytes"
